refactor(soyeon): 패키지 수정
