+++ src/main/java/sopt/study/testcode/soyeon/TestcodeApplication.java
@@ -1,4 +1,4 @@
-package sopt.study.testcode.spring;
+package sopt.study.testcode.soyeon;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
