+++ src/main/java/sopt/study/testcode/soyeon/unit/CafeKioskRunner.java
@@ -1,9 +1,9 @@
-package sopt.study.testcode.unit;
+package sopt.study.testcode.soyeon.unit;
 
 import java.time.LocalDateTime;
-import sopt.study.testcode.unit.beverage.Americano;
-import sopt.study.testcode.unit.beverage.Latte;
-import sopt.study.testcode.unit.order.Order;
+import sopt.study.testcode.soyeon.unit.beverage.Americano;
+import sopt.study.testcode.soyeon.unit.beverage.Latte;
+import sopt.study.testcode.soyeon.unit.order.Order;
 
 public class CafeKioskRunner {
