+++ src/main/java/sopt/study/testcode/soyeon/unit/order/Order.java
@@ -1,10 +1,10 @@
-package sopt.study.testcode.unit.order;
+package sopt.study.testcode.soyeon.unit.order;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import sopt.study.testcode.unit.beverage.Beverage;
+import sopt.study.testcode.soyeon.unit.beverage.Beverage;
 
 @Getter
 @RequiredArgsConstructor
